Only checks project references to be resolved
Currently we check for all the references to be resolved while getting P2P references, which adds an overhead in terms of UI delays since it will then need to resolve all the references (project or assembly).
This PR restrict this resolution only to project references, since that's what we use to generate the full closure of the project to restore.

Fixes https://github.com/NuGet/Home/issues/6120